requirements.txt
# GUI and Application
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0

# Audio and Text-to-Speech
gTTS>=2.3.1
pyttsx3>=2.90
playsound==1.2.2
pyaudio>=0.2.13
numpy>=1.24.0

# Hardware Communication
pyserial>=3.5
RPi.GPIO>=0.7.1  # For Raspberry Pi deployments
adafruit-circuitpython-dht==3.7.0
adafruit-circuitpython-bmp3xx==1.3.7
gps==3.19

# Data Management
pandas>=1.5.0
matplotlib>=3.6.0

# Utilities
python-dateutil>=2.8.2
requests>=2.28.1
pytz>=2022.7

# Development
pytest>=7.3.1
black>=22.12.0
flake8>=6.0.0
mypy>=1.0.0
